--- tests/Extension/Php70/InstanceTest.php
@@ -1,26 +1,0 @@
-<?php
-
-/*
- * This file is part of the Zephir.
- *
- * (c) Phalcon Team <team@zephir-lang.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Extension\Php70;
-
-use PHPUnit\Framework\TestCase;
-use Stub\Instance;
-
-class InstanceTest extends TestCase
-{
-    public function testShouldThrowErrorOnInstantiateAbstractClass(): void
-    {
-        $this->expectException(\Error::class);
-        $this->expectExceptionMessage('Cannot instantiate abstract class TestAbstractClass');
-
-        Instance::testInstanceCreate(\TestAbstractClass::class);
-    }
-}
